refactor: consolidate duplicate response helpers into lib/api-utils
